Donate CPU: Fix wildcard import anti-pattern
Using wildcard imports is very discouraged. While it could clutter
namespace and such things it is also a real problem for static
analyzers. I read about this and can confirm the issues.
So instead of wildcard imports, explicit imports of only the used things
are used now.
For details see for example
https://docs.quantifiedcode.com/python-anti-patterns/maintainability/from_module_import_all_used.html
and the references linked there.