Fix processing of for loop iteration expression in FwdAnalysis
The for loop iteration expression was processed following the flow of
token as well as during the looping.
It should not be processed with the flow of token, but only during the
looping.

This commit adds the facility to parse `for (InitStatement; Condition; IterationExpression)`
and skip the iteration expression.

This fixes false positives, for instance, in the added test, we used to
have the following warning:
```
test.cpp:4:style:Redundant initialization for 'i'. The initialized value is overwritten before it is read.\n
test.cpp:3:note:i is initialized\n
test.cpp:4:note:i is overwritten\n
```