Fix #9574: htmlreport: Defect ID column not scrollable anymore
This removes the fixed position setting for the menu on the left. Thus
it scrolls up and down like the rest of the page. The disadvantage is
that the menu is not available everywhere on the page, but IMHO it is
more important to have a working page where nothing overlaps or is
unreachable.