cppcheckdata.py: Reorder addon error message parts
When viewing the error output stream from addons being able to easily
find the error key (addon+errorID) is helpful.

Currently the error key is show after any description of the error
message. Since the error description is much more variable in length
than the error location it is difficult to visually parse the output,
especially if the description is long enough for the message to cause
the terminal to wrap the line.

Move the error key location to immediately after the error location.
This makes it much easier to see when looking at the console output.