testrunner: Allow to specify testcases with regex
Allow to specify which tests to run using a regex. In addition to the
currently existing ways of running tests, this makes it possible to
specify the tests to run using a regex. For example (depending on your
shell, you might have to quote the wildcard expressions):

Run all Memleak tests (TestMemleak, TestMemleakInFunction, etc)

	./testrunner TestMemleak.*

Run all simplifyOperatorName tests in TestTokenizer:

	./testrunner TestTokenizer::simplifyOperatorName.*

All tests containing 'const' or 'Const'

	make testrunner && ./testrunner .*::.*[C|c]onst.*