Fix false positive AssignmentIntegerToAddress
Due to a issue while parsing the function declaration (we ended up with
the type of the first argument instead of the return type).

```
c.cpp:6:11: portability: Assigning an integer to a pointer is not portable. [AssignmentIntegerToAddress]
  Array a = f(i);
          ^
```
typedef std::array<int, 2> Array;

Array f(int i);

int g(int i) {
  Array a = f(i);
  return 2;
}
```