Lazily compute ErrorMessage::hash
Since the computation is quite expensive, only do it if we need it. This
uses the assumptions that tokenList and the tokens in it are still valid
at the time of computation.
I believe this is the case, but I might be wrong.

As the hash is not used very often, this is very beneficial. On a huge
codebase with many warnings (typically many missingOverride on some
protobuf generated code), the cppcheck duration is halved, and I believe
calculateWarningHash is never called.