10110: Fix FP knownConditionTrueFalse
With
```
enum FENUMS { NONE = 0, CB = 8 };

bool calc(int flag) {
    if (!flag) {
        return false;
    }

    if (flag & CB) {
        return true;
    }
    return false;
}
```

Since flag is not a bool, we can't know its value in the second if.
However, isOppositeCond would still return true (since `!flag` and
`flag` really are opposite condition), and we would update flag with a
known value of `1`. This patch adds an additional check to prevent that
when we don't deal with a bool.