fix hangs from poor type lookup of base classes in derived classes
Cppcheck is inconsistent in handling the lookoup of uninstantiated
template types and their scopes.  Sometimes the template parameters are
taken into account and sometimes they are ignored. This can cause a
circular dependency when a template type is derived from itself with
different parameters.

This patch prevents an incorrect type lookup from causing a hang. I does
not fix the underlying problem of how uninstantiated template types are
represented and looked up.  That requires auditing a lot of code and
determining the best way to address this problem.