Fix UB on CTU attribute read.
Running on some project with UBSan enabled and using CTU results in a report for an integer overflow in `atoi` while reading back files written by cppcheck.

Values are written to to the file as `MathLib::bigint` (`long long`) and may fall out-of-range of `int`. `atoi` has UB if the read value is not representable. Use a function without UB and correct type instead. This still does not check the validity of the read value, though.

UBSan report observed with musl and probably not possible to observe with glibc, since glibc seems to call `strtoi` for `atoi` internally.