Fix false positive invalidFunctionArg for zero arguments to calloc.
The following program results in a false positive invalid argument
report for the first argument of calloc:

void f() { auto x = calloc(0,0); }

But calloc has (implementation-)defined behavior analoguesly to
malloc/realloc if either argument is zero. Modify the valid range
for the first argument to match that of the second and the other
allocation functions.