Support function level try blocks
Currently Cppcheck emits "syntax error: keyword 'try' is not allowed in
global scope" on valid code like this:
    void foo() try {} catch (int) {}
and skips the whole file where such code was found.

When function level try blocks are used with all functions except
constructors and destructors they act as if the function body consisted
of the following try-catch block. So additional simplification step is
added that performs this transformation.