donate_cpu_lip.get_cppcheck: early return without checkout if local repo is up to date
Modify function get_cppcheck to first fetch changes remote repository, then act depend on the result: 
* No new commit : local repository is up to date, return early. This skips checkout, meaning we do not revert the local cppcheck binary we just compiled. 
* New commit : to a checkout + merge. This situation is equivalent to the existing one, since pull is equivalent to fetch+merge (cf https://git-scm.com/docs/git-pull)

This should significantly improve DACA performance. 

Currently, for each package, we do a pull + checkout which is terribly inefficient because checkout causes the following "make" to fully re-compile and re-linke cppcheck, even if there is no new commit (and no code change). 
With this change, the following make will be a NoOp (make consider cppcheck up to date) if there is no new commit. But the following make would still be a full rebuild if there is a new commit.