addons/namingng: rename config file to namingng.json
Confusingly, the config file for namingng.py was named naming.json, while
another addon naming.py can be used using a *different* file that is also named
naming.json, residing in a different directory.

By renaming the config file to namingng.json, it is made explicit that this
file belongs to the namingng.py addon.