addons/namingng.py: Add namingng.json and namingng.config.json.
Confusingly, the example config file for namingng.py was named naming.json, while
another addon naming.py can be used using a *different* file that is also named
naming.json, residing in a different directory.

The feature where an addon can be named addon.json, referring with its "script"
key to the actual addon script, further adds to possible confusion.

By renaming the default namingng config file from naming.json to
namingng.config.json, it is made explicit that this file is a config file for
the namingng.py addon.

A file namingng.json is added as an example file to be used using
--addon=namingng.json.

This matches the setup of namingng as created and tested in test-other.py.