lib/errorlogger.cpp: Improve conversion from InternalError to ErrorMessage
ErrorMessage::setmsg() assumes that the supplied message string is a
combination of a short message and a verbose message, separated by a newline.

However, InternalError does not use that format, and the errorMessage of
InternalError may contain an unintentional newline. Then, when converting from
InternalError to ErrorMessage, the short message will receive the first line,
and the verbose message will receive the lines that follow.

This approach may lead to obscure messages during verbose logging since the
informative first, often explaining the reason for the InternalError line is
lost.

This is solved by setting the verbose message to the complete original error
string.

A concrete example is the failure of an addon to produce output starting with
'{', which is reported as an InternalError. The clarity of this message is
also improved with this patch, as the "Failure to execute" criterion was not
clear from the message. Also, the failing output in this case is moved to the
details field of InternalError.

Unit tests were added for multi-line InternalError conversions.