addons/namingng.py: Reinstate dict-with-regexps feature.
In 7c168fa850b774a4d62643e4c005f7f12d053587 (Feature/ros naming check, Jan 9,
2019) a feature was implemented where regular expressions can be given in a
dict, i.e. `{"[a-z]+":[arg0,arg1]}` instead of an array `["[a-z]+"]`.

There is only the namingng.py code, an example `ROS_naming.json` and some
GitHub discussion to go at when trying to understand the feature. The feature
was not readily understandable by looking only at the code. It tested the first
argument using `if ~arg0` (i.e. using bit-wise negation), which only evaluates
as true when `arg0 == -1`, while the example JSON and discussion indicate that
`arg0` is actually boolean. The code seemed to indicate some form where `arg1`
is not used, and `arg0` would contain a string, but it was not obvious how this
should work.

This feature seems to be intended as follows: When specifying a list, a
positive match of each item is required. When specifying a dict, the dict keys
are regular expressions and their values are `[bool,string]` with bool
indicating what match result is to be reported as an error, and string a
message that is appended to the reported error. This approach has various
benefits: the config can specify an explanatory message, and it is now possible
to more simply forbid and/or require certain patterns.

Recent changes, especially 98b9244bcbd92afd2a1d2acbec669e7783857cd0
(addons/namingng.py: Add tests for include guards, config file validation,
Jan 2 2024) completely broke this feature, while keeping the code in. This
patch provides a rewrite of the feature, adds sanity checks to the
configuration, and provides unit tests for both the configuration of the
feature and the feature itself.

The dict feature does not work for `RE_FILE` (yet). The example config file
`ROS_naming.json` is modified to prevent an error triggered by the dict value
of `RE_FILE`.