addons/namingng.py: Add unit test for standalone mode.
This reuses the existing unit test, running the addon on the dump file created.

The format of the standalone mode is different so the expected output is
slightly transformed before checking.

This implicitly also tests `cppcheck --dump`, i.e. whether checking 1 file
leads to 1 dump file, and whether the dump filename is indeed file+'.dump'.
This is relevant as some code simply assumes this to be the case.