addons/namingng.py: Implement backward compatibility with addons/naming.py.
Merging naming.py into namingng.py reduces redundant code and leverages
argument validation and the unit tests for namingng.py, with the aim to improve
maintainability.

naming.py supported RE_CONST to check constants, which is now a feature of
namingng.py as well; the unit test is expanded to cover the feature.

naming.py has been updated to call through to namingng.py.

Suppressions are updated to suppres the new namingng error id
[naming-namingConvention].

The unit tests for naming.py are kept as-is.