add test/cli/test-json.py: Schema validation for JSON files.
test-json.py uses jsonschema (https://pypi.org/project/jsonschema/) to validate
JSON files against schemas, optionally with custom format checkers.

All JSON files in the project must be accounted for in
test-json.schema_mapping(), either as a schema, or as a file to be validated
against a schema. The JSON schemas are implicitly validated against the most
recent metaschema available in jsonschema.

This commit adds schemas for JSON addon files and for namingng config files.

TODO:
- runtime schema validation
- remove current ad-hoc validation
- generate documentation from schemas